+++ client/src/components/CustomEvents/index.js
@@ -1,0 +1,3 @@
+import CustomEvents from './CustomEvents';
+
+export default CustomEvents;