+++ client/src/components/ServiceDetailsForm/index.js
@@ -1,0 +1,3 @@
+import ServiceDetailsForm from './ServiceDetailsForm';
+
+export default ServiceDetailsForm;