+++ server/routers/service.js
@@ -3,7 +3,7 @@ import serviceController from '../controllers/serviceController';
 
 const router = express.Router();
 
-router.post('/register', serviceController.registerService, (req, res) => {
+router.post('/', serviceController.registerService, (req, res) => {
   console.log('register service router is working');
   console.log(res.locals.response);
   res.status(200).json(res.locals.response);
@@ -15,7 +15,13 @@ router.get('/:id', serviceController.findServiceById, (req, res) => {
   res.status(200).json(res.locals.response);
 });
 
-router.put('/update', serviceController.updateServiceById, (req, res) => {
+router.put('/:id', serviceController.updateServiceById, (req, res) => {
+  console.log('update service router is working');
+  console.log(res.locals.response);
+  res.status(200).json(res.locals.response);
+});
+
+router.delete('/:id/:projectId', serviceController.deleteServiceById, (req, res) => {
   console.log('update service router is working');
   console.log(res.locals.response);
   res.status(200).json(res.locals.response);